Cherrypicked Commit - (PHP 7, 8) Fix memory leaks in curl and error handling (#1350)
* Fix memory leak when setting http response code in response to non-fatal error

Signed-off-by: Bob Weinand <bob.weinand@datadoghq.com>

* Add memory leak checking to webserver tests

* Fix curl handles leaking

There were two different leaks:

a) on PHP 7 the empty hashtable was not removed from the multi-map when headers were injected and removed
b) when curl handles were not closed (PHP 7 and 8) the curl handle simply leaked. This was especially bad on PHP 7 as resource handles are not reused. (On PHP 8 curl handles are an objects, whose handle is reused and implicitly bounds the maximum possible lost memory.)

While on PHP 7, there is not much choice, but to track curl resource destruction via an injected objects destructor, on PHP 8 we make use of the weakmaps mechanism.

Signed-off-by: Bob Weinand <bob.weinand@datadoghq.com>

* Fix PHP 5 exception handling memory leaks

Signed-off-by: Bob Weinand <bob.weinand@datadoghq.com>

* Fix leaking MINIT allocated memory when ddtrace gets disabled at (first) RINIT

Signed-off-by: Bob Weinand <bob.weinand@datadoghq.com>