[profiling] Avoid crashing when pcntl_fork is called (#1745)
This replaces the global profiler object in the child process with
None, and sets profiling_enabled to false. This leaks some memory
at the expense of crashing less.

However, it's still possible to have issues, for instance if a thread
in the process holds a lock at the time of the fork. So there is
further work to either synchronize such that no locks are held, or to
clean up all that stuff before the fork, and make it again after.

* Bump minor version

* Add args to run-tests to show error info

* Run x-profiling tests on PHP 8.1 too